requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
Pillow==10.2.0
python-dotenv==1.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-decouple==3.8
scipy==1.11.1